Customer Interface
Tentative user interface that takes input data from the user on their duration of stay and then outputs the results of a query text file and display them for the user. In the next update the select button will bring up a button and scroll input number to select the room. Still requires the SQL query mysql connector to generate the resulting text file from the user inputs.